+++ src/const.js
@@ -2,15 +2,12 @@ const DEFAULT_FILTER = `all`;
 
 const MAX_DESCRIPTION_LENGTH = 140;
 
-const GENRE = [
-  `Musical`,
-  `Drama`,
-  `Mystery`,
-  `Comedy`,
-  `Western`,
-  `Cartoon`,
-  `Film-Noir`,
-];
+const FILTER = new Map([
+  [`all`, `All movies`],
+  [`watchlist`, `Watchlist`],
+  [`history`, `History`],
+  [`favorites`, `Favorites`],
+]);
 
 const EMOJI = new Map([
   [`smile`, `smile.png`],
@@ -19,8 +16,5 @@ const EMOJI = new Map([
   [`angry`, `angry.png`],
 ]);
 
-const AGE = [`0`, `6`, `12`, `16`, `18`];
 
-const COUNTRY = [`BRA`, `IND`, `CAN`, `CHN`, `RUS`, `USA`, `TUR`, `FRA`, `JPN`];
-
-export {DEFAULT_FILTER, MAX_DESCRIPTION_LENGTH, GENRE, EMOJI, AGE, COUNTRY};
+export {DEFAULT_FILTER, FILTER, MAX_DESCRIPTION_LENGTH, EMOJI};
