--- src/mock/filter.js
@@ -1,9 +1,0 @@
-const Filter = new Map([
-  [`all`, `All movies`],
-  [`watchlist`, `Watchlist`],
-  [`history`, `History`],
-  [`favorites`, `Favorites`],
-]);
-
-
-export {Filter};
